feat: six-menu-items suffix/prefix can be used in six-select
