+++ docs/examples/docs-demo-six-select-10.vue
@@ -1,26 +1,16 @@
 <template>
 <div>
 
-        <six-select placeholder="Small" size="small" multiple>
+        <six-select placeholder="Select one">
+          <six-menu-label>Group 1</six-menu-label>
           <six-menu-item value="option-1">Option 1</six-menu-item>
           <six-menu-item value="option-2">Option 2</six-menu-item>
           <six-menu-item value="option-3">Option 3</six-menu-item>
-        </six-select>
-
-        <br>
-
-        <six-select placeholder="Medium" size="medium" multiple>
-          <six-menu-item value="option-1">Option 1</six-menu-item>
-          <six-menu-item value="option-2">Option 2</six-menu-item>
-          <six-menu-item value="option-3">Option 3</six-menu-item>
-        </six-select>
-
-        <br>
-
-        <six-select placeholder="Large" size="large" multiple>
-          <six-menu-item value="option-1">Option 1</six-menu-item>
-          <six-menu-item value="option-2">Option 2</six-menu-item>
-          <six-menu-item value="option-3">Option 3</six-menu-item>
+          <six-menu-divider></six-menu-divider>
+          <six-menu-label>Group 2</six-menu-label>
+          <six-menu-item value="option-4">Option 4</six-menu-item>
+          <six-menu-item value="option-5">Option 5</six-menu-item>
+          <six-menu-item value="option-6">Option 6</six-menu-item>
         </six-select>
       
 </div>
