+++ docs/examples/docs-demo-six-select-12.vue
@@ -1,11 +1,21 @@
 <template>
 <div>
 
-        <six-select label="Select one">
-          <six-menu-item value="option-1">Option 1</six-menu-item>
-          <six-menu-item value="option-2">Option 2</six-menu-item>
-          <six-menu-item value="option-3">Option 3</six-menu-item>
-        </six-select>
+        <div class="selecting-example">
+          <six-select placeholder="">
+            <six-menu-item value="option-1">Option 1</six-menu-item>
+            <six-menu-item value="option-2">Option 2</six-menu-item>
+            <six-menu-item value="option-3">Option 3</six-menu-item>
+          </six-select>
+
+          <br>
+
+          <six-button data-option="option-1">Set 1</six-button>
+          <six-button data-option="option-2">Set 2</six-button>
+          <six-button data-option="option-3">Set 3</six-button>
+        </div>
+
+        
       
 </div>
 </template>
@@ -15,6 +25,17 @@
 <script>
 export default {
   name: 'docs-demo-six-select-12',
-  mounted() {  }
+  mounted() { 
+          (() => {
+            const container = document.querySelector('.selecting-example');
+            const select = container.querySelector('six-select');
+
+            [...container.querySelectorAll('six-button')].map((button) => {
+              button.addEventListener('click', () => {
+                select.value = button.dataset.option;
+              });
+            });
+          })();
+         }
 }
 </script>
