+++ docs/examples/docs-demo-six-select-17.vue
@@ -1,10 +1,18 @@
 <template>
 <div>
 
-        <six-select id="async-select" async-filter filter-placeholder="Enter a number">
-          <six-menu-item id="async-menu-item" value="search_list_prompt">Use search to show entries</six-menu-item>
+        <six-select label="Experience" filter multiple select-all-button clearable>
+          <six-menu-item value="AUSTRALIA">Australia</six-menu-item>
+          <six-menu-item value="BRAZIL">Brazil</six-menu-item>
+          <six-menu-item value="CHINA">China</six-menu-item>
+          <six-menu-item value="EGYPT">Egypt</six-menu-item>
+          <six-menu-item value="FRANCE">France</six-menu-item>
+          <six-menu-item value="GERMANY">Germany</six-menu-item>
+          <six-menu-item value="INDIA">India</six-menu-item>
+          <six-menu-item value="JAPAN">Japan</six-menu-item>
+          <six-menu-item value="SPAIN">Spain</six-menu-item>
+          <six-menu-item value="UNITED_STATES">United States</six-menu-item>
         </six-select>
-        
       
 </div>
 </template>
@@ -14,40 +22,6 @@
 <script>
 export default {
   name: 'docs-demo-six-select-17',
-  mounted() { 
-          const asyncSelect = document.querySelector('#async-select');
-          const asyncMenu = document.querySelector('#async-menu-item').parentElement;
-
-          for (let i = 0; i < 500; i++) {
-            const child = document.createElement('six-menu-item');
-            child.innerText = `Value ${i}`;
-            child.value = `value-${i}`;
-            asyncMenu.appendChild(child);
-          }
-
-          const removeAllChildNodes = (parent) => {
-            while (parent.firstChild) {
-              parent.removeChild(parent.firstChild);
-            }
-          };
-
-          asyncSelect.addEventListener('six-async-filter-fired', ($event) => {
-            const filterValue = $event.detail.filterValue;
-            removeAllChildNodes(asyncMenu);
-            for (let i = 0; i < 500; i++) {
-              const child = document.createElement('six-menu-item');
-              child.innerText = `Value ${i}`;
-              child.value = `value-${i}`;
-              if (
-                filterValue.includes(`${i}`) ||
-                filterValue.toLocaleLowerCase().includes(`value`) ||
-                filterValue.toLocaleLowerCase().includes(`value ${i}`) ||
-                filterValue === ''
-              ) {
-                asyncMenu.appendChild(child);
-              }
-            }
-          });
-         }
+  mounted() {  }
 }
 </script>
