+++ docs/examples/docs-demo-six-select-20.vue
@@ -1,7 +1,7 @@
 <template>
 <div>
 
-        <six-select filter multiple id="six-select-dynamic-options"></six-select>
+        <six-select id="autocomplete-example" autocomplete clearable></six-select>
         
       
 </div>
@@ -13,11 +13,37 @@
 export default {
   name: 'docs-demo-six-select-20',
   mounted() { 
-          const sixSelectDynamicOptions = document.getElementById('six-select-dynamic-options');
-          sixSelectDynamicOptions.options = Array.from(Array(100).keys()).map((idx) => ({
-            label: `label ${idx}`,
-            value: `value ${idx}`,
-          }));
+          (() => {
+            const clearAllChildren = (node) => {
+              let child = node.lastElementChild;
+              while (child) {
+                node.removeChild(child);
+                child = node.lastElementChild;
+              }
+            };
+
+            const createMenuItem = (value, label) => {
+              const menuItem = document.createElement('six-menu-item');
+              menuItem.innerText = label;
+              menuItem.setAttribute('value', value);
+              return menuItem;
+            };
+
+            const select = document.getElementById('autocomplete-example');
+
+            select.addEventListener('six-select-change', (event) => {
+              if (event.detail.isSelected) {
+                // don't fetch new values on selection
+                return;
+              }
+              clearAllChildren(select);
+
+              const enteredText = event.detail.value || 'All Values';
+              new Array(5).fill('').forEach((item, idx) => {
+                select.append(createMenuItem(`option ${enteredText} ${idx}`, `Option ${enteredText} ${idx}`));
+              });
+            });
+          })();
          }
 }
 </script>
