+++ docs/examples/docs-demo-six-select-24.vue
@@ -1,8 +1,7 @@
 <template>
 <div>
 
-        <six-select id="multiple-error-text" label="List of string message" invalid></six-select>
-        
+        <six-select label="Simple string message" error-text="This is a simple string message" invalid> </six-select>
       
 </div>
 </template>
@@ -12,10 +11,6 @@
 <script>
 export default {
   name: 'docs-demo-six-select-24',
-  mounted() { 
-          const sixSelect = document.getElementById('multiple-error-text');
-          sixSelect.errorText = ['Message 1', 'Message 2'];
-          sixSelect.errorTextCount = 3;
-         }
+  mounted() {  }
 }
 </script>
