+++ docs/examples/docs-demo-six-select-8.vue
@@ -7,7 +7,7 @@
           <six-menu-item value="option-3">Option 3</six-menu-item>
           <six-menu-item value="option-4">Option 4</six-menu-item>
           <six-menu-item value="option-5">Option 5</six-menu-item>
-          <six-menu-item value="option-6">Option 6 (with tooltip due to long text)</six-menu-item>
+          <six-menu-item value="option-6">Option 6 (with tooltip due to long long long long text)</six-menu-item>
         </six-select>
       
 </div>
